refactor: merge document and project stores
