+++ src/main/ipc/index.ts
@@ -1,3 +1,3 @@
 export * from './context-menus';
 export * from './pdf';
-export * from './versioned-stores';
+export * from './project-stores';
