+++ src/modules/domain/project/adapters/git/git-project-store-manager/electron-renderer-ipc/index.ts
@@ -1,7 +1,6 @@
 import * as Effect from 'effect/Effect';
 import { pipe } from 'effect/Function';
 
-import { createAdapter as createVersionedDocumentStoreAdapter } from '../../../../../../../modules/domain/rich-text/adapters/versioned-document-store/git/electron-renderer-ipc-versioned-document-store';
 import { mapErrorTo } from '../../../../../../../utils/errors';
 import { RepositoryError as VersionedProjectRepositoryError } from '../../../../errors';
 import { type ProjectStoreManager } from '../../../../ports';
@@ -33,11 +32,10 @@ export const createAdapter = (): ProjectStoreManager => {
             mergeConflictInfo,
             remoteProjects,
           }) => ({
-            versionedProjectStore: createProjectStoreAdapter(),
-            versionedDocumentStore: createVersionedDocumentStoreAdapter({
-              projectId,
-              // It's really the main process store that manages the filesystem workdir here,
-              // but from the perspective of the client using this adapter it should be transparent.
+            // It's really the main process store that manages the filesystem
+            // workdir here, but from the perspective of the client using this
+            // adapter it should be transparent.
+            projectStore: createProjectStoreAdapter({
               managesFilesystemWorkdir: true,
             }),
             projectId,
@@ -74,11 +72,10 @@ export const createAdapter = (): ProjectStoreManager => {
             mergeConflictInfo,
             remoteProjects,
           }) => ({
-            versionedProjectStore: createProjectStoreAdapter(),
-            // It's really the main process store that manages the filesystem workdir here,
-            // but from the perspective of the client using this adapter it should be transparent.
-            versionedDocumentStore: createVersionedDocumentStoreAdapter({
-              projectId,
+            // It's really the main process store that manages the filesystem
+            // workdir here, but from the perspective of the client using this
+            // adapter it should be transparent.
+            projectStore: createProjectStoreAdapter({
               managesFilesystemWorkdir: true,
             }),
             projectId,
