--- src/modules/domain/project/commands/delete-document-from-project.ts
@@ -1,56 +1,0 @@
-import * as Effect from 'effect/Effect';
-import { pipe } from 'effect/Function';
-
-import {
-  NotFoundError as VersionedDocumentNotFoundError,
-  RepositoryError as VersionedDocumentRepositoryError,
-  ValidationError as VersionedDocumentValidationError,
-  type VersionedDocumentStore,
-} from '../../../../modules/domain/rich-text';
-import {
-  MigrationError,
-  type ResolvedArtifactId,
-} from '../../../../modules/infrastructure/version-control';
-import {
-  NotFoundError as VersionedProjectNotFoundError,
-  RepositoryError as VersionedProjectRepositoryError,
-  ValidationError as VersionedProjectValidationError,
-} from '../errors';
-import { type ProjectId } from '../models';
-import { type ProjectStore } from '../ports';
-
-export type DeleteDocumentFromProjectArgs = {
-  documentId: ResolvedArtifactId;
-  projectId: ProjectId;
-  deleteFromFilesystem?: boolean;
-};
-
-export type DeleteDocumentFromProjectDeps = {
-  deleteDocument: VersionedDocumentStore['deleteDocument'];
-  deleteDocumentFromProjectStore: ProjectStore['deleteDocumentFromProject'];
-};
-
-export const deleteDocumentFromProject =
-  ({
-    deleteDocument,
-    deleteDocumentFromProjectStore,
-  }: DeleteDocumentFromProjectDeps) =>
-  ({
-    documentId,
-    projectId,
-    deleteFromFilesystem,
-  }: DeleteDocumentFromProjectArgs): Effect.Effect<
-    void,
-    | VersionedProjectRepositoryError
-    | VersionedProjectNotFoundError
-    | VersionedProjectValidationError
-    | VersionedDocumentRepositoryError
-    | VersionedDocumentNotFoundError
-    | VersionedDocumentValidationError
-    | MigrationError,
-    never
-  > =>
-    pipe(
-      deleteDocumentFromProjectStore({ documentId, projectId }),
-      Effect.flatMap(() => deleteDocument({ documentId, deleteFromFilesystem }))
-    );
