--- src/modules/domain/rich-text/adapters/versioned-document-store/git/git-versioned-document-store/index.ts
@@ -1,798 +1,0 @@
-import * as Effect from 'effect/Effect';
-import { pipe } from 'effect/Function';
-import git, { type PromiseFsClient as IsoGitFsApi } from 'isomorphic-git';
-
-import {
-  type Branch,
-  type Change,
-  type ChangeId,
-  type Commit,
-  createGitBlobRef,
-  decomposeGitBlobRef,
-  getFileCommitHistory,
-  type GitBlobRef,
-  type GitCommitHash,
-  isGitBlobRef,
-  isGitCommitHash,
-  isUncommittedChangeId,
-  MigrationError,
-  parseBranch,
-  parseGitCommitHash,
-  parseGitCommitHashEffect,
-  readBlobAtCommit,
-  type ResolvedArtifactId,
-  type UncommitedChange,
-  UNCOMMITTED_CHANGE_ID,
-  VersionControlNotFoundErrorTag,
-  VersionControlRepositoryErrorTag,
-  versionedArtifactTypes,
-} from '../../../../../../../modules/infrastructure/version-control';
-import { fromNullable } from '../../../../../../../utils/effect';
-import { mapErrorTo } from '../../../../../../../utils/errors';
-import {
-  type Filesystem,
-  FilesystemNotFoundErrorTag,
-} from '../../../../../../infrastructure/filesystem';
-import {
-  DeletedDocumentError,
-  NotFoundError,
-  RepositoryError,
-  ValidationError,
-  VersionedDocumentDeletedDocumentErrorTag,
-  VersionedDocumentNotFoundErrorTag,
-} from '../../../../errors';
-import {
-  CURRENT_SCHEMA_VERSION,
-  PRIMARY_RICH_TEXT_REPRESENTATION,
-  type ResolvedDocument,
-  type RichTextDocument,
-} from '../../../../models';
-import { type VersionedDocumentStore } from '../../../../ports/versioned-document-store';
-
-export const createAdapter = ({
-  isoGitFs,
-  filesystem,
-  projectId: projId,
-  projectDir,
-  managesFilesystemWorkdir,
-}: {
-  // We have 2 filesystem APIs because isomorphic-git works well in both browser in Node.js
-  // with its own implemented fs APIs, which more or less comply to the Node.js API.
-  // In cases where we interact with the filesystem outside isomorphic-git (e.g. for listing files or normailizing paths),
-  // we are using our own Filesystem API.
-  isoGitFs: IsoGitFsApi;
-  filesystem: Filesystem;
-  projectId: string;
-  // projectDir is the repo root handed to isomorphic-git: it equals projectId on a real filesystem,
-  // but a virtual filesystem could root the repo elsewhere (e.g. '.').
-  projectDir: string;
-  managesFilesystemWorkdir: boolean;
-}): VersionedDocumentStore => {
-  // This is not an ideal model but we want to be able to tell that the document store we are searching in is the desired one.
-  // Without this we are risking registering interest in documents from other repositories (and therefore polluting our stores)
-  let projectId: string = projId;
-  const setProjectId: VersionedDocumentStore['setProjectId'] = (id) =>
-    Effect.sync(() => {
-      projectId = id;
-    });
-
-  const validateDocumentId: (
-    id: ResolvedArtifactId
-  ) => Effect.Effect<GitBlobRef, ValidationError, never> = (id) =>
-    pipe(
-      Effect.succeed(id),
-      Effect.filterOrFail(
-        isGitBlobRef,
-        (val) => new ValidationError(`Invalid document id: ${val}`)
-      )
-    );
-
-  const extractDocumentRelativePathFromId: (
-    id: ResolvedArtifactId
-  ) => Effect.Effect<string, ValidationError, never> = (id) =>
-    pipe(
-      validateDocumentId(id),
-      Effect.map((gitBlobRef) => {
-        const { path: documentPath } = decomposeGitBlobRef(gitBlobRef);
-        return documentPath;
-      })
-    );
-
-  const buildDocumentAbsolutePathFromId: (
-    id: ResolvedArtifactId
-  ) => Effect.Effect<string, ValidationError | RepositoryError, never> = (id) =>
-    pipe(
-      extractDocumentRelativePathFromId(id),
-      Effect.flatMap((relativeDocumentPath) =>
-        pipe(
-          filesystem.getAbsolutePath({
-            path: relativeDocumentPath,
-            dirPath: projectDir,
-          }),
-          Effect.catchAll(() =>
-            Effect.fail(new RepositoryError('Git repo error'))
-          )
-        )
-      )
-    );
-
-  const createDocument: VersionedDocumentStore['createDocument'] = ({
-    filePath,
-    writeToFile,
-    branch,
-  }) =>
-    pipe(
-      fromNullable(
-        filePath,
-        () =>
-          new ValidationError(
-            'File path is required when creating a document in the Git repo'
-          )
-      ),
-      Effect.flatMap((path) =>
-        pipe(
-          branch
-            ? Effect.succeed(branch)
-            : pipe(
-                getCurrentBranch(),
-                // Map errors related to branching to repo errors
-                Effect.catchAll(() =>
-                  Effect.fail(new RepositoryError('Git repo error'))
-                )
-              ),
-          Effect.flatMap((currentBranch) =>
-            Effect.try({
-              try: () =>
-                createGitBlobRef({
-                  ref: currentBranch,
-                  path,
-                }),
-              catch: mapErrorTo(
-                ValidationError,
-                'Cannot create the Git blob ref for the document'
-              ),
-            })
-          ),
-          Effect.tap(() =>
-            writeToFile
-              ? pipe(
-                  fromNullable(
-                    managesFilesystemWorkdir,
-                    () =>
-                      new RepositoryError(
-                        'This store does not manage a workdir'
-                      )
-                  ),
-                  Effect.flatMap(() =>
-                    pipe(
-                      filesystem.writeFile({ path, content: '' }),
-                      Effect.catchAll(() =>
-                        Effect.fail(new RepositoryError('Git repo error'))
-                      )
-                    )
-                  )
-                )
-              : Effect.succeed(undefined)
-          )
-        )
-      )
-    );
-
-  const findDocumentById: VersionedDocumentStore['findDocumentById'] = (id) =>
-    pipe(
-      buildDocumentAbsolutePathFromId(id),
-      Effect.flatMap((documentPath) =>
-        pipe(
-          filesystem.readTextFile(documentPath),
-          Effect.catchTag(FilesystemNotFoundErrorTag, () =>
-            Effect.fail(
-              new NotFoundError(`File with path ${documentPath} not found`)
-            )
-          ),
-          Effect.catchAll(() =>
-            Effect.fail(new RepositoryError('Git repo error'))
-          )
-        )
-      ),
-      Effect.map(
-        ({ content }) =>
-          ({
-            id,
-            artifact: {
-              type: versionedArtifactTypes.RICH_TEXT_DOCUMENT,
-              schemaVersion: CURRENT_SCHEMA_VERSION,
-              representation: PRIMARY_RICH_TEXT_REPRESENTATION,
-              content,
-            },
-            handle: null,
-          }) as ResolvedDocument
-      )
-    );
-
-  const isDocumentModified: (args: {
-    projectDir: string;
-    documentPath: string;
-  }) => Effect.Effect<boolean, RepositoryError | NotFoundError, never> = ({
-    projectDir,
-    documentPath,
-  }) =>
-    pipe(
-      Effect.tryPromise({
-        try: () =>
-          git.status({
-            fs: isoGitFs,
-            dir: projectDir,
-            filepath: documentPath,
-          }),
-        catch: mapErrorTo(RepositoryError, 'Git repo error'),
-      }),
-      Effect.flatMap((fileChangedStatus) => {
-        if (fileChangedStatus === 'ignored' || fileChangedStatus === 'absent') {
-          return Effect.fail(
-            new NotFoundError('Document is ignored or absent')
-          );
-        }
-
-        if (fileChangedStatus === 'unmodified') {
-          return Effect.succeed(false);
-        }
-
-        return Effect.succeed(true);
-      })
-    );
-
-  const getDocumentLastChangeId: VersionedDocumentStore['getDocumentLastChangeId'] =
-    (documentId) => {
-      const getLastModificationCommitId: (args: {
-        projectDir: string;
-        documentPath: string;
-      }) => Effect.Effect<GitCommitHash, RepositoryError, never> = ({
-        projectDir,
-        documentPath,
-      }) =>
-        pipe(
-          Effect.tryPromise({
-            try: () =>
-              git.log({
-                fs: isoGitFs,
-                dir: projectDir,
-                // get only the latest commit affecting the file
-                depth: 1,
-                filepath: documentPath,
-              }),
-            catch: mapErrorTo(RepositoryError, 'Git repo error'),
-          }),
-          Effect.flatMap((commits) =>
-            commits.length > 0
-              ? Effect.try({
-                  try: () => parseGitCommitHash(commits[0].oid),
-                  catch: mapErrorTo(RepositoryError, 'Git repo error'),
-                })
-              : Effect.fail(
-                  // TODO: Consider returning a NotFoundError
-                  // But since we know there is some kind of modification, we must get a commit that modified the document.
-                  new RepositoryError(
-                    'Could not find the last commit that modified the document'
-                  )
-                )
-          )
-        );
-
-      const getChangeId: (
-        modified: boolean
-      ) => (args: {
-        projectDir: string;
-        documentPath: string;
-      }) => Effect.Effect<ChangeId, RepositoryError, never> =
-        (modified) =>
-        ({ projectDir, documentPath }) =>
-          modified
-            ? Effect.succeed(UNCOMMITTED_CHANGE_ID)
-            : getLastModificationCommitId({ documentPath, projectDir });
-
-      return Effect.Do.pipe(
-        Effect.bind('documentPath', () =>
-          extractDocumentRelativePathFromId(documentId)
-        ),
-        Effect.flatMap(({ documentPath }) =>
-          pipe(
-            isDocumentModified({ projectDir, documentPath }),
-            Effect.flatMap((modified) =>
-              getChangeId(modified)({ projectDir, documentPath })
-            )
-          )
-        )
-      );
-    };
-
-  // Note: This function expects an absolute path in writeToFileWithPath.
-  // TODO: Express this accurately in the type system.
-  const updateRichTextDocumentContent: VersionedDocumentStore['updateRichTextDocumentContent'] =
-    ({ content, writeToFileWithPath }) =>
-      pipe(
-        fromNullable(
-          writeToFileWithPath,
-          () =>
-            new ValidationError(
-              'File path not provided; cannot write to document file'
-            )
-        ),
-        Effect.flatMap((path) => filesystem.writeFile({ path, content })),
-        Effect.catchAll(() =>
-          Effect.fail(new RepositoryError('Git repo error'))
-        )
-      );
-
-  const deleteDocument: VersionedDocumentStore['deleteDocument'] = ({
-    documentId,
-    deleteFromFilesystem,
-  }) =>
-    deleteFromFilesystem && managesFilesystemWorkdir
-      ? pipe(
-          buildDocumentAbsolutePathFromId(documentId),
-          Effect.flatMap((documentPath) =>
-            pipe(
-              filesystem.deleteFile({ path: documentPath }),
-              Effect.catchAll(() =>
-                Effect.fail(new RepositoryError('Git repo error'))
-              )
-            )
-          )
-        )
-      : Effect.succeed(undefined);
-
-  const getDocumentHistory: VersionedDocumentStore['getDocumentHistory'] = (
-    documentId
-  ) => {
-    const getDocumentCommitHistory = ({
-      projectDir: dir,
-      documentPath,
-    }: {
-      projectDir: string;
-      documentPath: string;
-    }): Effect.Effect<Commit[], RepositoryError, never> =>
-      pipe(
-        getFileCommitHistory({
-          isoGitFs,
-          dir,
-          filepath: documentPath,
-        }),
-        Effect.catchTag(VersionControlRepositoryErrorTag, (err) =>
-          Effect.fail(new RepositoryError(err.message))
-        )
-      );
-
-    const getDocumentChangeHistory: (args: {
-      modified: boolean;
-      commitHistory: Commit[];
-    }) => Effect.Effect<
-      {
-        history: Change[];
-        latestChange: Change;
-        lastCommit: Commit | null;
-        hasUncommittedChanges: boolean;
-      },
-      RepositoryError | NotFoundError,
-      never
-    > = ({ modified, commitHistory }) => {
-      const lastCommit = commitHistory.length > 0 ? commitHistory[0] : null;
-
-      if (modified) {
-        const uncommittedChange: UncommitedChange = {
-          id: UNCOMMITTED_CHANGE_ID,
-        };
-
-        return Effect.succeed({
-          history: [uncommittedChange, ...commitHistory],
-          latestChange: uncommittedChange,
-          lastCommit,
-          hasUncommittedChanges: true,
-        });
-      }
-
-      return Effect.succeed({
-        history: commitHistory,
-        // TODO: Handle more gracefully
-        latestChange: lastCommit!,
-        lastCommit,
-        hasUncommittedChanges: false,
-      });
-    };
-
-    return Effect.Do.pipe(
-      Effect.bind('documentPath', () =>
-        extractDocumentRelativePathFromId(documentId)
-      ),
-      Effect.bind('document', () => findDocumentById(documentId)),
-      Effect.flatMap(({ documentPath, document }) =>
-        Effect.Do.pipe(
-          Effect.bind('documentCommitHistory', () =>
-            getDocumentCommitHistory({ documentPath, projectDir })
-          ),
-          Effect.bind('isModified', () =>
-            isDocumentModified({ projectDir, documentPath })
-          ),
-          Effect.flatMap(({ documentCommitHistory, isModified }) =>
-            getDocumentChangeHistory({
-              modified: isModified,
-              commitHistory: documentCommitHistory,
-            })
-          ),
-          Effect.map(
-            ({ history, latestChange, lastCommit, hasUncommittedChanges }) => ({
-              history,
-              latestChange,
-              lastCommit,
-              current: document.artifact,
-              hasUncommittedChanges,
-            })
-          )
-        )
-      )
-    );
-  };
-
-  const getDocumentFromFs: (
-    documentId: ResolvedArtifactId
-  ) => Effect.Effect<
-    RichTextDocument,
-    ValidationError | RepositoryError | NotFoundError | MigrationError,
-    never
-  > = (documentId) =>
-    pipe(
-      findDocumentById(documentId),
-      Effect.map((resolvedDocument) => resolvedDocument.artifact)
-    );
-
-  const getDocumentAtCommit: (args: {
-    projectDir: string;
-    documentPath: string;
-    commitHash: GitCommitHash;
-  }) => Effect.Effect<
-    RichTextDocument,
-    RepositoryError | NotFoundError | DeletedDocumentError,
-    never
-  > = ({ projectDir, documentPath, commitHash }) => {
-    const readDocumentBlobAtCommit = (hash: GitCommitHash) =>
-      pipe(
-        readBlobAtCommit({
-          isoGitFs,
-          dir: projectDir,
-          commitHash: hash,
-          filepath: documentPath,
-        }),
-        Effect.mapError((error) =>
-          error._tag === VersionControlNotFoundErrorTag
-            ? new NotFoundError(
-                `Document "${documentPath}" not found at commit ${hash}`
-              )
-            : new RepositoryError('Git repo error')
-        )
-      );
-
-    return pipe(
-      readDocumentBlobAtCommit(commitHash),
-      // When the document is not found at this commit, check whether it
-      // existed in the parent commit. If it did, this is a deletion
-      // (DeletedDocumentError); otherwise the document never existed here
-      // (NotFoundError).
-      Effect.catchTag(VersionedDocumentNotFoundErrorTag, (notFoundError) =>
-        pipe(
-          Effect.tryPromise({
-            try: () =>
-              git.readCommit({
-                fs: isoGitFs,
-                dir: projectDir,
-                oid: commitHash,
-              }),
-            catch: () => notFoundError,
-          }),
-          Effect.flatMap((commitResult) => {
-            const parentOid = commitResult.commit.parent[0] ?? null;
-            if (!parentOid) return Effect.fail(notFoundError);
-
-            return pipe(
-              parseGitCommitHashEffect(parentOid),
-              Effect.flatMap(readDocumentBlobAtCommit),
-              // Document not in parent either — never existed here.
-              Effect.catchAll(() => Effect.fail(notFoundError)),
-              // Document exists in parent but not in this commit — deleted.
-              Effect.flatMap(() =>
-                Effect.fail(
-                  new DeletedDocumentError(notFoundError.message, {
-                    parentCommitId: parentOid,
-                  })
-                )
-              )
-            );
-          })
-        )
-      ),
-      Effect.flatMap((blob) =>
-        Effect.try({
-          try: () => Buffer.from(blob).toString('utf8'),
-          catch: mapErrorTo(RepositoryError, 'Git repo error'),
-        })
-      ),
-      Effect.map((content) => ({
-        type: versionedArtifactTypes.RICH_TEXT_DOCUMENT,
-        schemaVersion: CURRENT_SCHEMA_VERSION,
-        representation: PRIMARY_RICH_TEXT_REPRESENTATION,
-        content,
-      }))
-    );
-  };
-
-  const getDocumentAtChange: VersionedDocumentStore['getDocumentAtChange'] = ({
-    documentId,
-    changeId,
-  }) => {
-    return isUncommittedChangeId(changeId)
-      ? getDocumentFromFs(documentId)
-      : Effect.Do.pipe(
-          Effect.bind('documentPath', () =>
-            extractDocumentRelativePathFromId(documentId)
-          ),
-          Effect.flatMap(({ documentPath }) =>
-            pipe(
-              Effect.succeed(changeId),
-              Effect.filterOrFail(
-                isGitCommitHash,
-                (val) => new ValidationError(`Invalid commit hash: ${val}`)
-              ),
-              Effect.flatMap((commitHash) =>
-                getDocumentAtCommit({ projectDir, documentPath, commitHash })
-              )
-            )
-          )
-        );
-  };
-
-  const getUncommittedDocumentStateHash: (
-    documentId: ResolvedArtifactId
-  ) => Effect.Effect<
-    string,
-    ValidationError | RepositoryError | NotFoundError | MigrationError,
-    never
-  > = (documentId) =>
-    pipe(
-      getDocumentFromFs(documentId),
-      Effect.flatMap(({ content }) =>
-        Effect.tryPromise({
-          try: () =>
-            git.hashBlob({
-              object: content,
-            }),
-          catch: mapErrorTo(RepositoryError, 'Git repo error'),
-        })
-      ),
-      Effect.map((hashBlobResult) => hashBlobResult.oid)
-    );
-
-  const getDocumentHashAtCommit: (args: {
-    projectDir: string;
-    documentPath: string;
-    commitHash: GitCommitHash;
-  }) => Effect.Effect<string, RepositoryError, never> = ({
-    projectDir,
-    documentPath,
-    commitHash,
-  }) =>
-    pipe(
-      Effect.tryPromise({
-        try: () =>
-          git.resolveRef({ fs: isoGitFs, dir: projectDir, ref: commitHash }),
-        catch: mapErrorTo(RepositoryError, 'Git repo error'),
-      }),
-      Effect.flatMap((commitOid) =>
-        Effect.tryPromise({
-          try: () =>
-            git.readBlob({
-              fs: isoGitFs,
-              dir: projectDir,
-              oid: commitOid,
-              filepath: documentPath,
-            }),
-          catch: mapErrorTo(RepositoryError, 'Git repo error'),
-        })
-      ),
-      Effect.map((readBlobResult) => readBlobResult.oid)
-    );
-
-  const getDocumentHashAtChange: (args: {
-    documentId: ResolvedArtifactId;
-    changeId: Change['id'];
-  }) => Effect.Effect<
-    string,
-    ValidationError | RepositoryError | NotFoundError | MigrationError,
-    never
-  > = ({ documentId, changeId }) =>
-    isUncommittedChangeId(changeId)
-      ? getUncommittedDocumentStateHash(documentId)
-      : Effect.Do.pipe(
-          Effect.bind('documentPath', () =>
-            extractDocumentRelativePathFromId(documentId)
-          ),
-          Effect.bind('commitHash', () =>
-            pipe(
-              Effect.succeed(changeId),
-              Effect.filterOrFail(
-                isGitCommitHash,
-                (val) => new ValidationError(`Invalid commit hash: ${val}`)
-              )
-            )
-          ),
-          Effect.flatMap(({ documentPath, commitHash }) =>
-            getDocumentHashAtCommit({
-              projectDir,
-              documentPath,
-              commitHash,
-            })
-          )
-        );
-
-  const isContentSameAtChanges: VersionedDocumentStore['isContentSameAtChanges'] =
-    ({ documentId, change1, change2 }) => {
-      if (isUncommittedChangeId(change1) && isUncommittedChangeId(change2)) {
-        return Effect.succeed(true);
-      }
-
-      return Effect.Do.pipe(
-        Effect.bind('hash1', () =>
-          getDocumentHashAtChange({ documentId, changeId: change1 })
-        ),
-        Effect.bind('hash2', () =>
-          getDocumentHashAtChange({ documentId, changeId: change2 })
-        ),
-        Effect.map(({ hash1, hash2 }) => hash1 === hash2)
-      );
-    };
-
-  const discardUncommittedChanges: VersionedDocumentStore['discardUncommittedChanges'] =
-    ({ documentId, writeToFileWithPath }) =>
-      pipe(
-        getDocumentHistory(documentId),
-        Effect.flatMap(({ lastCommit, hasUncommittedChanges }) => {
-          if (!hasUncommittedChanges) {
-            return Effect.fail(
-              new NotFoundError(
-                'The document does not have uncommitted changes to discard.'
-              )
-            );
-          }
-
-          if (!lastCommit) {
-            return Effect.fail(
-              new RepositoryError(
-                'The document only has uncommitted changes (and no commits). Cannot restore to a known state.'
-              )
-            );
-          }
-
-          return pipe(
-            extractDocumentRelativePathFromId(documentId),
-            Effect.flatMap((documentPath) =>
-              pipe(
-                Effect.succeed(lastCommit.id),
-                Effect.filterOrFail(
-                  isGitCommitHash,
-                  (val) => new ValidationError(`Invalid commit hash: ${val}`)
-                ),
-                Effect.flatMap((commitHash) =>
-                  getDocumentAtCommit({
-                    projectDir,
-                    documentPath,
-                    commitHash,
-                  })
-                ),
-                // If the document was deleted in the last commit, restore
-                // from the parent commit instead.
-                Effect.catchTag(
-                  VersionedDocumentDeletedDocumentErrorTag,
-                  (e) =>
-                    e.data.parentCommitId
-                      ? pipe(
-                          getDocumentAtCommit({
-                            projectDir,
-                            documentPath,
-                            commitHash: parseGitCommitHash(
-                              e.data.parentCommitId
-                            ),
-                          }),
-                          // The parent commit itself may also not contain the
-                          // document (e.g. consecutive deletions). We don't
-                          // recurse further — treat this as unrecoverable.
-                          Effect.catchTag(
-                            VersionedDocumentDeletedDocumentErrorTag,
-                            (e) => Effect.fail(new RepositoryError(e.message))
-                          )
-                        )
-                      : Effect.fail(
-                          new RepositoryError(
-                            'Document was deleted but has no parent commit to restore from.'
-                          )
-                        )
-                )
-              )
-            ),
-            Effect.flatMap((documentAtCommit) =>
-              updateRichTextDocumentContent({
-                documentId,
-                representation: documentAtCommit.representation,
-                content: documentAtCommit.content,
-                writeToFileWithPath,
-              })
-            )
-          );
-        })
-      );
-
-  const getCurrentBranch: () => Effect.Effect<
-    Branch,
-    ValidationError | RepositoryError | NotFoundError,
-    never
-  > = () =>
-    pipe(
-      Effect.tryPromise({
-        try: () =>
-          git.currentBranch({
-            fs: isoGitFs,
-            dir: projectDir,
-          }),
-        catch: mapErrorTo(
-          RepositoryError,
-          'Error in getting the current branch'
-        ),
-      }),
-      Effect.filterOrFail(
-        (currentBranch) => typeof currentBranch === 'string',
-        () =>
-          new NotFoundError(
-            'Could not retrieve the current branch. The repo is in detached HEAD state.'
-          )
-      ),
-      Effect.flatMap((currentBranch) =>
-        Effect.try({
-          try: () => parseBranch(currentBranch),
-          catch: mapErrorTo(RepositoryError, 'Could not parse current branch'),
-        })
-      )
-    );
-
-  const resolveContentConflict: VersionedDocumentStore['resolveContentConflict'] =
-    ({ documentId }) =>
-      pipe(
-        extractDocumentRelativePathFromId(documentId),
-        Effect.flatMap((documentPath) =>
-          Effect.tryPromise({
-            try: () =>
-              git.add({
-                fs: isoGitFs,
-                dir: projectDir,
-                filepath: documentPath,
-              }),
-            catch: mapErrorTo(RepositoryError, 'Git repo error'),
-          })
-        )
-      );
-  // This is a no-op in the Git document repo.
-  const disconnect: VersionedDocumentStore['disconnect'] = () =>
-    Effect.succeed(undefined);
-
-  return {
-    projectId,
-    setProjectId,
-    managesFilesystemWorkdir,
-    createDocument,
-    findDocumentById,
-    getDocumentLastChangeId,
-    updateRichTextDocumentContent,
-    deleteDocument,
-    getDocumentHistory,
-    getDocumentAtChange,
-    isContentSameAtChanges,
-    discardUncommittedChanges,
-    resolveContentConflict,
-    disconnect,
-  };
-};
