--- src/modules/domain/rich-text/adapters/versioned-document-store/git/git-versioned-document-store/utils.ts
@@ -1,3 +1,0 @@
-export const isNodeError = (err: unknown): err is NodeJS.ErrnoException => {
-  return err instanceof Error && 'code' in err;
-};
