+++ src/modules/domain/rich-text/commands/index.ts
@@ -1,2 +1,1 @@
 export { suggestMerge } from './suggest-merge';
-export { processDocumentChange } from './process-document-change';
