+++ src/modules/domain/rich-text/index.ts
@@ -3,7 +3,6 @@ export * from './models/heading';
 export * from './models/link';
 export { createAdapter as createPandocDocumentAnalyzerAdapter } from './adapters/pandoc-document-analyzer';
 export { createAdapter as createPandocRepresentationTransformAdapter } from './adapters/pandoc-representation-transform';
-export { createAdapter as createGitDocumentStoreAdapter } from './adapters/versioned-document-store/git/git-versioned-document-store';
 export * as prosemirror from './prosemirror';
 export * from './models';
 export * from './ports';
