+++ src/modules/domain/rich-text/ports/index.ts
@@ -3,4 +3,3 @@ export * from './document-analyzer';
 export * from './merge-conflict-resolver';
 export * from './representation-transform';
 export * from './pdf-engine';
-export * from './versioned-document-store';
